添加后台管理

+++ src/main/java/com/meiyun/jkan/repository/LabelRepository.java
@@ -8,6 +8,6 @@ import com.meiyun.jkan.model.Label;
  * 标签数据中心
  * @author larry.qi
  */
-public interface LabelRepository extends JpaRepository<Label, Integer> {
+public interface LabelRepository extends JpaRepository<Label, Long> {
 
 }
